moved tests to /libraries/ directory
