+++ network-api/networkapi/wagtailpages/tests/libraries/research_hub/test_author_index.py
@@ -5,8 +5,8 @@ from wagtail_localize import synctree
 
 from networkapi.wagtailpages.factory import profiles as profiles_factory
 from networkapi.wagtailpages.factory.research_hub import relations as relations_factory
-from networkapi.wagtailpages.tests.research_hub import base as research_test_base
-from networkapi.wagtailpages.tests.research_hub import utils as research_test_utils
+from networkapi.wagtailpages.tests.libraries.research_hub import base as research_test_base
+from networkapi.wagtailpages.tests.libraries.research_hub import utils as research_test_utils
 
 
 class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
