+++ network-api/networkapi/wagtailpages/tests/libraries/research_hub/test_library_page.py
@@ -13,8 +13,8 @@ from networkapi.wagtailpages.factory.research_hub import (
     taxonomies as taxonomies_factory,
 )
 from networkapi.wagtailpages.pagemodels.libraries.research_hub import constants
-from networkapi.wagtailpages.tests.research_hub import base as research_test_base
-from networkapi.wagtailpages.tests.research_hub import utils as research_test_utils
+from networkapi.wagtailpages.tests.libraries.research_hub import base as research_test_base
+from networkapi.wagtailpages.tests.libraries.research_hub import utils as research_test_utils
 
 
 class TestResearchLibraryPage(research_test_base.ResearchHubTestCase):
